+++ src/TRegx/CleanRegex/Match/Offset/MatchOffsetLimit.php
@@ -1,6 +1,10 @@
 <?php
 namespace TRegx\CleanRegex\Match\Offset;
 
+use TRegx\CleanRegex\Exception\CleanRegex\Messages\NoFirstElementFluentMessage;
+use TRegx\CleanRegex\Internal\Factory\NotMatchedFluentOptionalWorker;
+use TRegx\CleanRegex\Match\FluentMatchPattern;
+
 class MatchOffsetLimit implements OffsetLimit
 {
     /** @var callable */
@@ -34,5 +38,10 @@ class MatchOffsetLimit implements OffsetLimit
     public function only(int $limit): array
     {
         return \call_user_func($this->allFactory, $limit, false);
+    }
+
+    public function fluent(): FluentMatchPattern
+    {
+        return new FluentMatchPattern($this->all(), new NotMatchedFluentOptionalWorker(new NoFirstElementFluentMessage()));
     }
 }