+++ test/Unit/TRegx/CleanRegex/Match/MatchPattern/group_offset/fluent/MatchPatternTest.php
@@ -1,0 +1,108 @@
+<?php
+namespace Test\Unit\TRegx\CleanRegex\Match\MatchPattern\group_offset\fluent;
+
+use InvalidArgumentException;
+use PHPUnit\Framework\TestCase;
+use TRegx\CleanRegex\Exception\CleanRegex\NonexistentGroupException;
+use TRegx\CleanRegex\Internal\InternalPattern;
+use TRegx\CleanRegex\Match\MatchPattern;
+
+class MatchPatternTest extends TestCase
+{
+    /**
+     * @test
+     */
+    public function shouldGet_groups()
+    {
+        // given
+        $pattern = new MatchPattern(InternalPattern::standard('(?<two>[A-Z][a-z])?(?<rest>[a-z]+)'), 'Nice Matching Pattern');
+
+        // when
+        $twoGroups = $pattern->group('two')->offsets()->fluent()->all();
+        $restGroups = $pattern->group('rest')->offsets()->fluent()->all();
+
+        // then
+        $this->assertEquals([0, 5, 14], $twoGroups);
+        $this->assertEquals([2, 7, 16], $restGroups);
+    }
+
+    /**
+     * @test
+     */
+    public function shouldGet_unmatchedGroups()
+    {
+        // given
+        $pattern = new MatchPattern(InternalPattern::standard('(?<hour>\d\d)?:(?<minute>\d\d)?'), 'First->11:__   Second->__:12   Third->13:32');
+
+        // when
+        $hours = $pattern->group('hour')->offsets()->fluent()->all();
+        $minutes = $pattern->group('minute')->offsets()->fluent()->all();
+
+        // then
+        $this->assertEquals([7, null, 38], $hours);
+        $this->assertEquals([null, 26, 41], $minutes);
+    }
+
+    /**
+     * @test
+     */
+    public function shouldReturnEmptyArray_onNotMatchedSubject()
+    {
+        // given
+        $pattern = new MatchPattern(InternalPattern::standard('(?<two>[A-Z][a-z])?(?<rest>[a-z]+)'), 'NOT MATCHING');
+
+        // when
+        $groups = $pattern->group('two')->offsets()->fluent()->all();
+
+        // then
+        $this->assertEquals([], $groups);
+    }
+
+    /**
+     * @test
+     */
+    public function shouldThrow_onNonExistentGroup()
+    {
+        // given
+        $pattern = new MatchPattern(InternalPattern::standard('(?<existing>[a-z]+)'), 'matching');
+
+        // then
+        $this->expectException(NonexistentGroupException::class);
+        $this->expectExceptionMessage("Nonexistent group: 'missing'");
+
+        // when
+        $pattern->group('missing')->offsets()->fluent()->all();
+    }
+
+    /**
+     * @test
+     */
+    public function shouldThrow_onNonExistentGroup_onNotMatchedSubject()
+    {
+        // given
+        $pattern = new MatchPattern(InternalPattern::standard('(?<existing>[a-z]+)'), 'NOT MATCHING');
+
+        // then
+        $this->expectException(NonexistentGroupException::class);
+        $this->expectExceptionMessage("Nonexistent group: 'missing'");
+
+        // when
+        $pattern->group('missing')->offsets()->fluent()->all();
+    }
+
+    /**
+     * @test
+     */
+    public function shouldThrow_onInvalidGroupName()
+    {
+        // given
+        $pattern = new MatchPattern(InternalPattern::standard('(?<existing>[a-z]+)'), 'matching');
+
+        // then
+        $this->expectException(InvalidArgumentException::class);
+        $this->expectExceptionMessage("Group name must be an alphanumeric string starting with a letter, given: '2invalid'");
+
+        // when
+        $pattern->group('2invalid')->offsets()->fluent()->all();
+    }
+}